Move category actions/reducer analog to account. Remove unused account LoadAction
