--- src/app/store/category.reducer.ts
@@ -1,13 +1,0 @@
-import { Category } from './model';
-import { ADD, ActionsTypes, CategoryActions } from './category.actions';
-
-const initialState: Category[] = [];
-
-export function categoryReducer(state: Category[] = initialState, action: ActionsTypes): Category[] {
-    switch (action.type) {
-        case ADD:
-            return [...state, action.payload];
-        default:
-            return state;
-    }
-}
